Agregue mas test, termine con los tramos

+++ src/sistema/Sistema.java
@@ -8,6 +8,7 @@ import estructuras.ArbolB;
 import estructuras.GrafoMatriz;
 import estructuras.Hash;
 import estructuras.HashPropiedad;
+import estructuras.Lista;
 import estructuras.Queue;
 import interfaces.ISistema;
 import sistema.Enumerados.Rubro;
@@ -27,6 +28,7 @@ public class Sistema implements ISistema {
 	//POST: Quedan todas las estructuras inicializadas para manejar el sistema (Hash, Arboles, etc).
 	public Retorno inicializarSistema (int cantPuntos) {
 		Retorno ret;
+		//Verificar si la cantidad de puntos ingresado es menor o igual a 0
 		if(cantPuntos <= 0){
 			ret = new Retorno(Resultado.ERROR_1);
 			return ret;
@@ -59,16 +61,19 @@ public class Sistema implements ISistema {
 	//POST: La esquina queda registrada en el sistema.
 	@Override
 	public Retorno registrarEsquina(double coordX, double coordY) {
+		//Verificar si hay lugar para ingresar la esquina
 		if(!matrizMapa.hayLugar()){
 			System.out.println("No hay lugar");
 			return new Retorno(Resultado.ERROR_1);
 		}
 		
+		//Verificar si la las coordenadas ya existen
 		if(tableHash.pertenece(coordX, coordY)){
 			System.out.println("hay uno igual");
 			return new Retorno(Resultado.ERROR_2);
 		}
 		
+		//Ingresar la esquina en el Hash general y en el grafo
 		Esquina e = new Esquina(coordX, coordY);
 		int pos = tableHash.insertar(e);
 		matrizMapa.agregarVertice(pos);
@@ -82,20 +87,24 @@ public class Sistema implements ISistema {
 	//POST: El punto de interes queda registrado en el sistema.
 	@Override
 	public Retorno registrarPuntoInteres(double coordX, double coordY, Rubro rubro, String nombre) {
+		//Verificar si hay lugar para ingresar la esquina
 		if(!matrizMapa.hayLugar()){
 			System.out.println("No hay lugar");
 			return new Retorno(Resultado.ERROR_1);
 		}
 		
+		//Verificar si el nombre ingresado es nulo o vacío
 		if(nombre == null || nombre == ""){
 			return new Retorno(Resultado.ERROR_2);
 		}
 		
+		//Verificar si la las coordenadas ya existen
 		if(tableHash.pertenece(coordX, coordY)){
 			System.out.println("hay uno igual");
 			return new Retorno (Resultado.ERROR_3);
 		}
 		
+		//Ingresar el Punto de Interes en el Hash general y en el grafo
 		PuntoDeInteres pi = new PuntoDeInteres(coordX, coordY, rubro, nombre);
 		int pos = tableHash.insertar(pi);
 		matrizMapa.agregarVertice(pos);
@@ -104,41 +113,48 @@ public class Sistema implements ISistema {
 	}
 
 	//PRE: Alguna celda del Hash = null || celda del hash = (0.0,0.0).
-		//El valor de direccion != null || direccion!= “” (vacio) y no exista en el sistema
+		//El valor de direccion != null || direccion != “” (vacio) y no exista en el sistema
 		//Hay al menos un vendedor registrado
 		//coordX y coordY no existe en el sistema.
 	//POST: La propiedad queda registrada en el sistema.
 	@Override
 	public Retorno registrarPropiedad(double coordX, double coordY, TipoPropiedad tipoPropiedad, String direccion) {
+		//no hay lugar
 		if(!matrizMapa.hayLugar()){
 			return new Retorno(Resultado.ERROR_1);
 		}
-		
+		//si la direccion es null o vacia
 		if(direccion == null || direccion == ""){
 			return new Retorno(Resultado.ERROR_2);
 		}
-		
-		//Falta error 3 si no existen vendedores registrados!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-		
+		//si no hay vendedores registrados
+		if(this.arbolDeVendedores.esArbolVacio()){
+			return new Retorno(Resultado.ERROR_3);
+		}
+		//si las coordenadas ya están siendo usadas
 		if(tableHash.pertenece(coordX, coordY)){
 			System.out.println("hay uno igual");
 			return new Retorno (Resultado.ERROR_4);
 		}
-		
-		if(tableHashProp.existeDireccion(direccion)){
+		//si la dirección ya existe
+		if(tableHashProp.pertenece(direccion)){
 			System.out.println("Error 5");
 			return new Retorno(Resultado.ERROR_5);
 		}
-		
-		
-		//Falta asignarle el vendedor!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-		
-
 		//Se agrega la propiedad al Hash general
 		Propiedad p = new Propiedad(coordX, coordY, tipoPropiedad, direccion);
-
 		int pos = tableHash.insertar(p);
 		matrizMapa.agregarVertice(pos);
+		//se asigna al has de propiedades del sistema
+		this.tableHashProp.insertar(p);
+		
+		//se le asigna al vendedor
+		Vendedor vendAsignado = (Vendedor)this.queueDeVendedores.front();
+		System.out.println("nombre vendedor asignado " +vendAsignado.getNombre());
+		vendAsignado.getHashPropiedades().insertar(p);
+		//lo saco del frente de la queue y lo mando al final
+		this.queueDeVendedores.dequeue();
+		this.queueDeVendedores.enqueue(vendAsignado);
 		
 		return new Retorno(Resultado.OK);
 	}
@@ -170,21 +186,29 @@ public class Sistema implements ISistema {
 	//POST:  El tramo queda registrado en el sistema
 	@Override
 	public Retorno registrarTramo(double coordXi, double coordYi, double coordXf, double coordYf, int peso) {
+		//Verificar si el peso es menor o igual a 0
 		if(peso <= 0){
 			return new Retorno(Resultado.ERROR_1);
 		}
 		
+		//Verificar si las coordenadas del tramo no existen en el sistema
 		if(!tableHash.pertenece(coordXi, coordYi) || !tableHash.pertenece(coordXf, coordYf)){
+			System.out.println("no existen las coordenadas en el sistema");
 			return new Retorno (Resultado.ERROR_2);
 		}
-		
-		//Si existe el tramo en el sistema error3
-		
+
         int origen = tableHash.devolverPosActual(coordXi, coordYi);
         int destino = tableHash.devolverPosActual(coordXf, coordYf);
+        
+        //Verificar si existe el tramo en el grafo
+		if(matrizMapa.existeArista(origen, destino, coordXi, coordYi, coordXf, coordYf)){
+			System.out.println("ya existe el tramo en el grafo");
+			return new Retorno (Resultado.ERROR_3);
+		}
 
-		matrizMapa.agregarArista(origen, destino, peso);
-		matrizMapa.agregarArista(destino, origen, peso);
+		//Agregar arista al grafo, como es no dirigido se ingresa 2 veces cambiando origen y destino
+		matrizMapa.agregarArista(origen, destino, peso, coordXi, coordYi, coordXf, coordYf);
+		matrizMapa.agregarArista(destino, origen, peso, coordXf, coordYf, coordXi, coordYi);
 			
 		return new Retorno(Resultado.OK);
 	}
@@ -195,14 +219,20 @@ public class Sistema implements ISistema {
 			return new Retorno (Resultado.ERROR_1);
 		}
 		
-		//Si existe el tramo en el sistema error2
-		
 		int origen = tableHash.devolverPosActual(coordXi, coordYi);
         int destino = tableHash.devolverPosActual(coordXf, coordYf);
         
+		//Verificar si no existe el tramo en el grafo
+		if(!matrizMapa.existeArista(origen, destino, coordXi, coordYi, coordXf, coordYf)){
+			System.out.println("no existe el tramo en el grafo");
+			return new Retorno (Resultado.ERROR_2);
+		}
+				
+		//Eliminar arista del grafo
 		matrizMapa.eliminarArista(origen, destino);
 		matrizMapa.eliminarArista(destino, origen);
-        
+
+		System.out.println("Exito");
 		return new Retorno(Resultado.OK);
 	}
 
@@ -256,13 +286,49 @@ public class Sistema implements ISistema {
 
 	@Override
 	public Retorno puntoInteresMasCercano(String direccionPropiedad, Rubro rubroPuntoInteres) {
-		//TODO reemplazar por su implementacion
+		//propiedad de esa direccion no existe 
+		if(this.tableHashProp.pertenece(direccionPropiedad)){
+			return new Retorno(Resultado.ERROR_1);
+		}
+		//si no hay ningun punto de interes de ese rubro
+		if(!this.tableHash.existePuntoDeEseRubro(rubroPuntoInteres)){
+			return new Retorno(Resultado.ERROR_2);
+		}
+		//si hay algún punto pero no es alcanzable
+		Propiedad p = this.tableHashProp.devolverPropiedad(direccionPropiedad);
+		int keyATableHash = 0;
+		if(p != null){
+			keyATableHash = this.tableHash.h(p.getCoordX(), p.getCoordY());
+		}
+		Lista verticesAdyascentes = this.matrizMapa.obtenerVerticesAdyacentes(keyATableHash);
+		if(!verticesAdyascentes.chequearSiAlMenosUnoDeRubro(rubroPuntoInteres)){
+			return new Retorno(Resultado.ERROR_3);
+		}
+		//TODO mostrar el pto de interes más cercano
 		return new Retorno();
 	}
 
 	@Override
 	public Retorno caminoMinimo(String direccionPropiedad, Double coordX, Double coordY) {
-		//TODO reemplazar por su implementacion
+		//esa direccion no esta registrada en el sistema
+		if(!this.tableHashProp.pertenece(direccionPropiedad)){
+			return new Retorno(Resultado.ERROR_1);
+		}
+		//ambas coordenadas no estan registradas
+		if(!this.tableHash.pertenece(coordX, coordY)){
+			return new Retorno(Resultado.ERROR_2);
+		}
+		//no hay camino posible entre prop y pto de interes
+		Propiedad p = this.tableHashProp.devolverPropiedad(direccionPropiedad);
+		int keyDePropEnTableHash = 0;
+		if(p != null){
+			keyDePropEnTableHash = this.tableHash.h(p.getCoordX(), p.getCoordY());
+		}
+		int keyDePtoInteresEnTableHash = this.tableHash.h(coordX, coordY);
+		if(this.matrizMapa.sonAdyacentes(keyDePropEnTableHash, keyDePtoInteresEnTableHash)){
+			return new Retorno(Resultado.ERROR_3);
+		}
+		//TODO mostrar el camino mínimo a pto de interes
 		return new Retorno();
 	}
 
@@ -278,7 +344,11 @@ public class Sistema implements ISistema {
 			return new Retorno(Resultado.ERROR_2);
 		}
 		//si no listar propiedades
-		return new Retorno(Resultado.OK, v.listarPropiedades());
+		System.out.println("es vacio " + v.getCedula() + v.getHashPropiedades().esVacio());
+		String retorno = "";
+		retorno = v.listarPropiedadesDelVendedor();
+		System.out.println();
+		return new Retorno(Resultado.OK, retorno);
 	}
